api/requirements.txt
fastapi
uvicorn
playwright
google-generativeai
beautifulsoup4
python-dotenv
lxml
sparticuz-chromium==123.0.0